// src/services/types/admin/admin-search-admins.ts
/** biome-ignore-all lint/correctness/noUnusedImports: IDONCARE */
/** biome-ignore-all lint/suspicious/noExplicitAny: IDONCARE */
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 * Source: api-docs
 * Title: CMS API
 * Description: CMS API
 * OpenAPI spec version: v1.0.0
 */

import type {ApiResultOnePageAdminSummaryResponse} from "../api-result-one-page-admin-summary-response.ts";

export const adminSearchAdminsQueryParamsSearchOptionEnum = {LOGIN_ID: "LOGIN_ID", NAME: "NAME"} as const;

export type AdminSearchAdminsQueryParamsSearchOptionEnumKey =
	(typeof adminSearchAdminsQueryParamsSearchOptionEnum)[keyof typeof adminSearchAdminsQueryParamsSearchOptionEnum];

export type AdminSearchAdminsQueryParams = {
	/**
	 * @description 권한 유형
	 * @type integer | undefined, int64
	 */
	permissionGroupId?: number;
	/**
	 * @description 삭제 여부
	 * @type boolean | undefined
	 */
	removed?: boolean;
	/**
	 * @description 잠금일시 검색 시작일
	 * @type string | undefined, date
	 */
	lockedStartAt?: string;
	/**
	 * @description 잠금일시 검색 종료일
	 * @type string | undefined, date
	 */
	lockedEndAt?: string;
	/**
	 * @description 검색 옵션(LOGIN_ID, NAME)
	 * @type string | undefined
	 */
	searchOption?: AdminSearchAdminsQueryParamsSearchOptionEnumKey;
	/**
	 * @description 검색어
	 * @type string | undefined
	 */
	searchKeyword?: string;
	/**
	 * @description 조회 대상 필드 목록
	 * @type array | undefined
	 */
	field?: string[];
	/**
	 * @description 조회 연산자 목록
	 * @type array | undefined
	 */
	operator?: string[];
	/**
	 * @description 조회 값 목록
	 * @type array | undefined
	 */
	value?: string[];
	/**
	 * @type integer | undefined, int32
	 */
	page?: number;
	/**
	 * @type integer | undefined, int32
	 */
	size?: number;
};

/**
 * @description OK
 */
export type AdminSearchAdmins200 = ApiResultOnePageAdminSummaryResponse;

export type AdminSearchAdminsQueryResponse = AdminSearchAdmins200;

export type AdminSearchAdminsQuery = {Response: AdminSearchAdmins200; QueryParams: AdminSearchAdminsQueryParams; Errors: any};